requirements.txt
anthropic>=0.40.0
groq>=0.9.0
langgraph>=0.2.50
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-docx>=1.1.2
python-dotenv>=1.0.0
pydantic>=2.9.0
gradio>=4.0.0

# Search provider — swap SEARCH_PROVIDER env var to change backend
tavily-python>=0.3.0
